+++ packages/manager/apps/dedicated/client/app/dedicated/server/dashboard/index.js
@@ -6,7 +6,12 @@ import '@uirouter/angularjs';
 import 'angular-translate';
 import '@ovh-ux/ui-kit';
 
+import ovhManagerBilling from '@ovh-ux/manager-billing';
 import advancedFeatures from './advanced-features';
+import commitment from './commitment';
+import cancelCommitment from './cancel-commitment';
+import cancelResiliation from './cancel-resiliation';
+import resiliation from './resiliation';
 
 import component from './dashboard.component';
 import routing from './dashboard.routing';
@@ -16,10 +21,15 @@ const moduleName = 'ovhManagerDedicatedServerDashboard';
 angular
   .module(moduleName, [
     advancedFeatures,
+    commitment,
+    cancelCommitment,
+    cancelResiliation,
     atInternet,
     'oui',
     'pascalprecht.translate',
     'ui.router',
+    ovhManagerBilling,
+    resiliation,
   ])
   .component('dedicatedServerDashboard', component)
   .config(routing)