+++ packages/manager/modules/models/index.js
@@ -1,10 +1,13 @@
 import Bill from './billing/Bill.class';
 import BillingService from './BillingService/BillingService.class';
+import CatalogPricing from './catalog/CatalogPricing.class';
 import Certificate from './certificate/Certificate.class';
+import Commitment from './service/Commitment.class';
 import DebtAccount from './billing/DebtAccount.class';
 import DedicatedCloud from './dedicatedCloud/dedicatedCloud.class';
 import PartnerLevel from './support/level/PartnerLevel.class';
 import Price from './Price/Price.class';
+import Pricing from './service/Pricing.class';
 import Service from './service/Service.class';
 import SupportLevel from './support/level/SupportLevel.class';
 import Ticket from './support/ticket/ticket.class';
@@ -13,11 +16,14 @@ import User from './user/User.class';
 export {
   Bill,
   BillingService,
+  CatalogPricing,
   Certificate,
+  Commitment,
   DebtAccount,
   DedicatedCloud,
   PartnerLevel,
   Price,
+  Pricing,
   Service,
   SupportLevel,
   Ticket,
@@ -27,7 +33,9 @@ export {
 export default {
   Bill,
   BillingService,
+  CatalogPricing,
   Certificate,
+  Commitment,
   DebtAccount,
   DedicatedCloud,
   PartnerLevel,